requirements.txt
certifi==2018.10.15
chardet==3.0.4
Click==7.0
dominate==2.3.4
Flask==1.0.2
Flask-Assets==0.12
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.2
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
python-dotenv==0.9.1
requests==2.20.1
urllib3==1.24.1
visitor==0.1.3
webassets==0.12.1
Werkzeug==0.14.1
WTForms==2.2.1
